Copy over task settings in AWS for the ongoing DMS task
It looks like AWS generates extra task settings for the DMS task.
As a result, when terrafor, compares the `task settings.json` with the task settings in AWS, it would plan an update-in-place/modification on the existing task.

This commit copies over the task settings as they are in the AWS console so that terraform doesn't detect any changes and therefore trigger any modifications.